Quebra processo.py em vários componentes
