+++ escavador/__init__.py
@@ -2,6 +2,7 @@ import escavador.v1
 import escavador.v2
 
 from ratelimit import RateLimitException
+from escavador.exceptions import FailedRequest
 
 from escavador.v1 import (
     Busca,
