+++ escavador/v2/__init__.py
@@ -1,1 +1,4 @@
+from .resources.envolvido import Envolvido
+from .resources.movimentacao import Movimentacao
 from .resources.processo import Processo
+from .resources.tribunal import Tribunal
